requirements.txt
streamlit==0.82.0
requests==2.25.1
Pillow==8.2.0